fix(skills): extract shared deep-freeze helpers and fix tuple recursion
- New deep_freeze.py module centralizes freeze/thaw logic (was duplicated
  between skill.py and skill_example.py)
- _freeze_any now recurses into tuple elements, not just lists
- ExtractionEngine._build_examples thaws values before str() serialization
  so frozen tuples render as lists in prompts
- _thaw return type simplified to Any
